OvmfPkg: VirtNorFlash: Move Low-Level NOR Flash Functions into a Library
This patch refactors low-level NOR flash functions into a dedicated library
to enable reuse, particularly for supporting the firmware upgrade feature.
